backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain
langchain-community
langchain-google-genai
langchain-text-splitters
chromadb
pypdf
pydantic
 